Deprecate unicode domain names for SRVConnector.
Author: ralphm
Reviewer: tom.prince
Fixes: #6245

twisted.names.dns.Names was recently changed to reject unicode domain names,
breaking XMPP clients passing the domain part of a JID to SRVConnector. This
change issues a deprecation warning if a unicode domain name is passed and
attempts to convert the domain name to an all-ASCII byte string for backwards
compatibility.

Note that, to support Internationalized Domain Names (IDN), applications must
convert the domain name to its ASCII Compatible Equivalent (ACE) as specified
in RFC 3490. The result can then be used as the domain to connect to with
SRVConnector.


git-svn-id: svn://svn.twistedmatrix.com/svn/Twisted/trunk@36867 bbbe8e31-12d6-0310-92fd-ac37d47ddeeb