Merge sshendpoint-default-ui-6550
Author: exarkun
Reviewer: therve
Fixes: #6550

When `None` (the default) is given for the `ui` parameter to
`twisted.conch.endpoints.SSHCommandClientEndpoint.newConnection`,
use a `ConsoleUI` which will read from `/dev/tty`.  If `/dev/tty`
is not accessible, use a non-UI object that answers no to all
prompts.

This replaces the old behavior of failing when trying to use `None`
as though it were a UI object.


git-svn-id: svn://svn.twistedmatrix.com/svn/Twisted/trunk@39095 bbbe8e31-12d6-0310-92fd-ac37d47ddeeb