Merge deferredlist-cancellation-6639: Added cancel() method to DeferredList.
Author: kaizhang
Reviewer: itamar
Fixes: #6639

Added cancellation support to twisted.internet.defer.DeferredList. When a DeferredList is cancelled by calling the cancel method, cancel every Deferred in the list. If the DeferredList has already fired, including the case where the fireOnOneCallback/fireOnOneErrback flag is set and the DeferredList fires because one Deferred in the list fires with a non-failure/failure result, do nothing in the cancel method.


git-svn-id: svn://svn.twistedmatrix.com/svn/Twisted/trunk@39657 bbbe8e31-12d6-0310-92fd-ac37d47ddeeb