Merge pop3-imports-6761: Correct twisted.mail.pop3's __all__
Author: shira
Reviewer: lvh
Fixes: #6761

Several POP3 client related classes were documented as being part of the API of the wrong module.

This potentially breaks backwards compatibility for star imports of a module that explicitly told you not to use those classes from there, which was deemed an acceptable bending of the compatibility policy (by lvh and itamar).


git-svn-id: svn://svn.twistedmatrix.com/svn/Twisted/trunk@40568 bbbe8e31-12d6-0310-92fd-ac37d47ddeeb